Fully metric reconstruction, regular and VR paths.
Lest I forget, the scale params are hard-coded, but here they are again:

g_fOBJZMetricMult = 44.72
g_fOBJGlobalMetricMult = 1.432

The only difference in the regular and VR paths is that the Y axis is inverted
because SteamVR requires a slightly different coord sys.

A lot more work is needed to integrate this with the actual VR code; but I think
this is a good start. This reconstruction matches the OPTs almost perfectly.