SBSVertexShader uses Metric VR reconstruction.
Added plumbing for a new CB at register b6 to store all the constants needed
for a full metric reconstruction and updated the backprojection code in
SBSVertexShader to use this instead.

A few effects are slightly broken at the moment, but the reconstruction looks
good. Also, cockpits need a wider FOV that matches the headset's FOV. At the
moment, I'm using "xwahacker_fov = 100" in the DC files to work around this; but
in the future the FOV can be computed automatically from the raw camera matrix
params returned by SteamVR.