projectMetric validated. Sun tagging uses Metric coords. Detection improved.
The projectMetric code has been validated to work in non-VR mode. The external
HUD shader displays the right position of the light (it matches the Sun Centroid)

The Sun Detection code is better now since it can detect when a light is visible
on the screen by using projectMetric().

Shadow Mapping needs to be checked in VR mode.