Fixed XWA Light tagging in VR path. Made object depth metric.
XWA lights are now tagged properly in VR mode. I had to force the SunCentroids
to be computed in metric space all the time and added a flag in projectMetric
to force the non-VR path even in VR mode. It works, though.

Also modified the SBSVertexShader back-projection to make the depth of the
various elements metric (this has yet to be confirmed).