Fixed the Active Cockpit. QPC code (SteamVR). POVOffsets switch.
The Active Cockpit now works in Metric mode in both VR and non-VR paths.

Added QPC code to selectively call WaitGetPoses. It does improve the framerate
in the mirror screen, but in the VR headset, the last presented frame is shown
as "stuck". It is easy to see around the edges and while it may be refreshing
faster, it doesn't feel right (feels a bit like the loading screen: stuck
at a higher framerate but lower refresh -- if that makes sense). Instead, I
think something must be presented all the time. The code is disabled by default.

Added an "apply_custom_POVs" switch to selectively apply the POVOffsets.