Fixed FOVscale, y_center for all resolutions and aspect ratios.
The metric reconstruction needs the raw FOVscale and y_center; but the shaders
need fixed FOVscale and y_center. This algorithm seems to work fine for a number
of resolutions and aspect ratios. The cutoff point between raw/fixed FOVscale,
y_center seems to be at an aspect ratio of 1.5996 (1.2 * 1.333).