First Draft: y_center auto-adjusted on a per-craft basis.
We're capturing the reticle center (crosshairs) limits in in-game coords and
comparing the centroid of that box with the position of the center of the
screen (after it has been compensated for cockpit camera pitch). Then, we're
using the difference between the reticle centroid and the screen center to
compute a proper y_center that will vary depending on the current cockpit.

This is semi-automatic for the moment, but it seems to work when
PreserveAspectRatio = 1. Still lots of testing needed, but this seems like a
good start.