Plumbing and draft for reticle resize in VR.
The reticle is now captured to an offscreen buffer if VR is on. The texture is
then sent to the ExternalHUDShader and displayed in that shader in red.

The reticle center is also captured on every frame, but it will only fix the
y_center if the cockpit is displayed and y_center hasn't been computed. We'll
need the reticle center on every frame so that we know what's the center of the
scale transform that will be applied later.