CMD fades to noise, fixed reticle y_center computation.
When a craft is destroyed, the CMD will fade to noise in cockpit view to hide
the edges created by the explosion. Also added a cool noise effect.

The y_center isn't computed properly 100% of the time: sometimes reticle.y is 0
when pitch is 0 (looking straight ahead). I added checks to prevent this
scenario. Also added code to compute y_center if Focal_Length.cfg can't be
loaded.