Improve SteamVR perf relocating WaitGetPoses()
With WaitGetPoses() at the end of Flip(), the CPU get's blocked while
the compositor waits for the GPU to finish the postprocessing work.

To allow the CPU to continue its work for the next frame, we delay
the sync point until we start sending GPU work for rendering the 2 eye
views.