Enable HMD rotation in 2D menus + Tech Room (Draft).
This commit enables yaw,pitch,roll tracking while 2D content is being displayed,
including the Tech Room. It is now possible to have a big 2D screen and look
around to the edges.

There's an artifact when the pause menu is presented: it does not clear the
previous buffer so it accumulates in the back. That has to be fixed now.