Fixed HUD and Reticle Aspect Ratio in SteamVR mode.
When using the Oculus' per-eye resolution of 1344x1600, we get some very ugly
vertical squashing. This commit fixes the HUD and reticle, but the sun flare
shader still has to be fixed.

I had to add more compensation params to the MetricRecCB to fix the reticle and
HUD distortion in the various modes.

Changed the default Zoom out scale for the HUD and fixed the HUD zoom out
function.

Added new DEFAULT_GUI_SCALE for the HUD so that it's easier to see without
looking all over the place.

Added "steamvr_mirror_window_scale" VRParam to control how zoomed in is the view
in the mirror window.

Added "steamvr_yaw_pitch_roll_from_mouse_look" to use mouse look to look around
the HUD in SteamVR mode. Also refactored the code that reads keyboard input to
apply yaw and pitch in the 2D menus.

Got rid of g_fGUIElemPZThreshold. Replaced it with "apply_uv_comp" since I need
that in the SBSVertexShader to apply the aspect ratio compensation only when
rendering the HUD.

Did some cleanup on the various Build*VertexBuffer functions.

Made the triangle pointer a bit sharper.