Partial fix for the reticle aspect ratio and sun flares in SBS mode.
There's some debug code in the ExternalHUDShader, so this should be removed
before publishing this. However, the aspect ratio of the reticle and the sun
flares is finally correct.

To get this to work, the desktop must be set to an integer multiple of the HMD's
resolution. For the PSVR, use multiples of 1920x1080 like 2880x1620 or 3000x1687