Holographic DC elems.
Removed hard edges by rendering the hologram as a plastic material with no
glossiness. The text is shadeless, though, to improve readability. Still needs
more tweaking as it doesn't look that much as a Star Wars hologram... but it
works.

Also added some functions to save and restore the blend state.