Keys can be used again to adjust the FOV on DC-enabled cockpits.
I broke the FOV adjustment through the keyboard on DC-enabled cockpits. The vars
that control the FOV do not increase with the keys, so they have to be modified
as well (only the global FOV was being updated).

Also refactored the name of the global FOV type so that it makes more sense.