Holdo bug fix.
Turns out it's possible to crash into things right after activating the hyperdrive.
If this happens, we return to the hangar but the hyperspace effect remains in
effect. This "Holdo maneuver effect" can be fixed by resetting the hyperspace
FSM if we're in the hangar.

This is a necessary change to fix this "Holdo bug", but more changes may be
needed as well.