Tweaks to how DC displays the HUD.
The HUD can now be toggled by pressing the D key. The initial state of the HUD
is controlled by a new global setting called "HUD_visible_on_startup" in
DynamicCockpit.cfg. If set to 1, the HUD will be visible (the "erase_region"
commands will be ignored). If set to 0, the "erase_region" commands will be
applied, erasing the HUD. Each time a mission is loaded, the visibility of the
HUD is set by this flag again.