Tweaks to the hologram effect.
I'm using the third channel of ssMask for shadelessness, this way, the holograms
can stay shadeless and I can increase the shadelessness gradually. This was
necessary to make the background of the holograms stay the same color regardless
of the normals behind them.

Also made the banding less obvious and tried to make the holograms more readable.