Fixes to the cycling the FOV setting.
When loading the FOVs from DC files, we ended up overwriting the xwahacker FOV
with the xwahacker large FOV because SetCurrentShipFOV does that to save the
settings when increasing the FOVs with the hotkeys. The fix is to prevent
SetCurrentShipFOV from overwriting these values when loading them from the DC
files.