More fixes to the FOV. The 1920x1080 adjustment should only be applied on startup.
Instead of adjusting the FOV for 1920x1080 every time the FOV is changed, let's
only adjust it when a cockpit is loaded. Otherwise, this constant adjustment
will effectively negate the effect of the FOV increase/decrease hotkeys (this
was very easy to see when setting the in-game resolution to 1280x1024, for
instance).