Additional text can now be rendered. Updated CraftInstanceHardpoint.
Piggy-backing on top of Jeremy's new Text Renderer, it is now possible to display
custom messages on the screen. There's a commented example in PrimarySurface::FLip
right before the call to this->RenderText();

Also added the CraftInstanceHardpoint struct along with some explanations on
how it works. This struct defines lasers as well as warheads and is used along
with PlayerDataTable.NumWarheadLauncherGroups, and NumberOfLasers.

I think it should now be possible to render DC text from first principles (i.e.
the data contained in PlayerDataTable and its links) instead of just capturing
mis-aligned text.