Missiles and Countermeasures can now be cleared and replaced.
Refactored the code that renders new text into RenderText(). Added a new block
that can erase specific DC elements (missiles and countermeasures) so that their
contents can be replaced with new text. The new text still has to be produced
from the XWA heap, but that's doable (at the moment it's just dumb placeholders
like "CC-CC" and "XX-XX")