Made hangar shadows darker to avoid transparency issues.
I'm not sure the hangar shadows are 100% fixed, but I've noticed some slight
transparency issues in the standard implementation as well. I think the proper
way to fix this is to render all shadows to a separate buffer as fully-black
areas and then blend them together, but that's going to require a lot more effort
and I'm not sure it's worth it.