Added new DC text lines taken from korekwerner. Enabled transparent DC elems.
I just took KW's definitions for the top, close-to-the-top and CMD text areas and
added them as new areas. I also enabled full transparency in the DC shader so
that we can move the text out of the way (some cockpits conflict with the text).

We can also put this text lines in our cockpits now. They might be a bit mis-aligned,
but it doesn't seem to be that bad after all.