Using barycentric coords to attach lights to laser textures.
Instead of looking for (1,1) in the uv coords for a laser texture, let's use
uv coords specified in the mat files and look for those coords in the current
texture. I'm using a variant of "ComputeCentroid" here to see if a texture
contains a specific uv coord, and then using barycentric coords to get the
reconstructed 3D coord that is added to the list.

The new mat property looks like this:

[Tex00000,TEX00000]
Light = 0.0, 1.0, 0.0
light_uv_coord_pos = 0.1, 0.9  <---

If this isn't specified, (0.1, 0.9) is used by default.