Fix memory bug in RenderBracket when shutting down.
Sometimes, when shutting down, an exception is raised. When running in debug mode,
this problem pointed back to the ID2D1SolidColorBrush. By a process of
trial-and-error the problem was narrowed down to the RenderBracket function. In
this function, I added two new brushes to redirect the output to either the
offscreen or DC RTVs. One of these brushes had to be set back to nullptr to avoid
leaving a dangling reference to one of these brushes when ddraw was unloaded.

Or something like that anyway.

The problem seems to be fixed now.