Added AlphaToBloomPS and PixelShaderNoGlass, plus lots of DS2 mat files.
The AlphaToBloomPS is used to map alpha to bloom. It's activated with the
"AlphaToBloom = 1" setting in the mat file. This helps apply bloom to textures
that have alpha, but no illumination texture.

If "AlphaToBloom = 1" and "NoColorAlpha = 1" are both set, then every component
in the output is zeroed-out except for the bloom. This makes an effect that
glows and looks blurry, like an energy field. This is used in DSReactorCylinder.

AlphaToBloomPS uses a new setting: "EffectBloom" to control how much bloom is
applied.

I finally added a pixel shader that does *not* convert semi-transparent textures
to glass materials. It's a simplified version of PixelShaderTexture. This shader
is activated by setting "AlphaIsntGlass = 1" in the mat file. If this shader is
enabled and "Ambient" is in the mat file, then textures get an "ambient" value
(faked by using the shadeless component). The "Ambient" setting is only used in
the PixelShaderNoGlass shader for now. I guess I can probably add it to
PixelShaderTexture as well.

Also added a bunch of material files for DS2 to fix how stuff looks in the last
mission.