Lava speed is now consistent all the time. Lava tiling is less obvious.
The lava shader was modified to do less repetitions in the tiling path (the uvs
were being needlessly multiplied by 3.0 before). Minor code cleanup as well.

I'm now using the Performance Counters to count the milliseconds between each
frame so that the lava animation speed stays consistent regardless of the
framerate. This technique should also be used to animate other things like the
sub-component elements in the CMD, for instance.