Split texture tags for smoke, explosions and electricity.
Moved smoke texture resnames from the Explosion list down to the Smoke list.

Renamed the Explosion resnames and boolean property to Electricity: it now
consists of only sparks and electric arcs.

Made a new resnames list with the explosion animations only, but changed the
code so that the final effect doesn't change (bloom parameters are applied to
either electricity or explosion tags).

This refactor is a necessary step to merge the new explosion shader with the
explosion textures.