Adding SharedMem to get data from the CockpitLook hook.
Since the hooks are loaded first, here we're opening an existing shared memory
object in the Local namespace. The object has two fields: a boolean that tells
us when the data is valid and a void* that can be cast to anything we want.
Since the hooks and ddraw share the heap, we only really need to read this
pointer once as long as it points to an area in the heap.