Wireframe CMD color bug: currentTargetIndex can be 0
I think I remember seeing the game crash if I let currentTargetIndex be 0 when
getting the color for the wireframe display; but Jeremy just told me that 0 is
a valid index, and I didn't see things crashing either. For an example of the
bug see mission 7, battle 3 and target Red 1. This can also be seen in any
skirmish mission when targeting the first craft.