The DAT GroupId and ImageID are now cached the first time they are parsed.
Not sure if this is the best way to put this code, but it works. Also added
bHasMaterial as a shortcut for checking if there is a texture *and* if it has
a material.

Manually cherry-picked from: 9100c8f35bd1c52e1c91d4885dd9a75f5f7f60e3