Make procedural explosions use the proper bloom setting. Fix animation.
The explosions_strength setting in Bloom.cfg was not being used in the
ExplosionShader. That's now fixed.

When using only the procedural explosions, the animation was a bit off: it
started empty and then the explosion appears. This was fixed by moving the range
of ExplosionTime from 4..-1 to 3..-1.

Manually cherry-picked from: 1eda68938e4ca987acfe6dab73a5d01636010c64