Hyperspace Tunnel speed fix and Hi-Res perf counters infrastructure.
Turns out that the explosions leave the hyperspace tunnel speed at -1 after they
are rendered, causing the hyper tunnel to move slowly backwards. We need to save
this setting on its own and restore the hyper tunnel speed when we render it.

Also tried to refactor the high res counters to control the lava shader speed,
but it doesn't work very well. I need to go back to that later.