Render Laser/Ion Energy levels from first principles.
This feature is to fix the Lambda Shuttle: it has 8 lasers, but one of them
doesn't display. So, instead, I'm introducing an algorithm to render the energy
levels from first principles. These levels are rendered on a new DC region:

EIGHT_LASERS_BOTH_SRC_IDX

; Slot 40: Eight lasers, Lambda Shuttle
source_def = 256,128,  60,-30, 199,1

More work is needed, but this is a good start.