Transparency Fix: New PixelShaderAnimLightMap.
The transparency used in regular light maps (as the game generates them) doesn't
blend very well around the edges because the alpha component has to be
multiplied by 10 and we also increase the Value in HSV mode. Animated Light maps
don't need their alphas to be multiplied by 10, and we need to saturate() the
output from the HSV conversion.