Fix crash due to early access into the CraftInstance table.
It looks like sometimes the craftInstance table/pointer is not valid during the
first few frames of a new mission. I've seen some intermittent cores when
playing with the GUN or the MIS and then switching to other craft. So adding
g_iPresentCounter > 5 seems to prevent this problem, since we wait for a few
frames at the beginning of the mission before we try to access the CraftInstance
table.