Enabled LightMap animations for DC cover textures added frame_rand.
DC cover textures can now have animated lightmaps. The lightmap is usually not
rendered because it interferes with the display, but if the texture is illuminated
and has an associated lightmap animation, then it will be displayed.

"frame_rand" was added to randomize a "frame_seq" animation.

g_GameEvent is now reset at the beginning of each mission.