frame_seq can specify single frame, introduced AlphaIsBloomMask.
If the alpha mode in frame_seq is set to 2, it will cause the alpha channel to
mask the bloom component. This can be used to display bloom only through "alpha
holes" punched on textures (but it requires that the regular texture displays
electric arcs, for instance, while the lightmap texture displays the "cover
texture with the alpha hole punched in).