Adding support for real-time POVOffset.
The current POV can be moved using Shift+Arrow Keys. The adjusted POV is
automatically saved to the relevant Ship.ini file and loaded the next time the
ship is loaded. This POV is mostly for VR, but it also works in regular non-VR
mode.

The POVOffset is written to a shared memory singleton so that the CockpitLook
hook can read it and apply it.