Refactored Normal Mapping code, moved bits to allow 3 greebles. Optimized code.
I moved all the normal mapping code to its own include file, so that we can use
it elsewhere. Moved some of the bits of the GreebleControl, so that we can have
3 greeble textures and 1 control bit to specify normal mapping. Added some
checks to avoid computing greebles if they won't be visible. Made the first
greeble texture optional.

This code runs a bit faster than the previous version as it will skip normal
mapping cotangent computations if they are too far away.