Added UVDisp and UVDisp+NormalMapping Blending Modes.
Refactored the Normal Mapping code so that we can compute the cotangent frame
once, even if multiple normal mapping modes are specified.

Added UVDisp and UVDisp+NormalMapping Blending Modes. UVDisp will use the first
two channels of a normal map as UV displacement, so that pixelation is broken.

UVDisp+NormalMapping does the same as the previous mode, but it also applies
normal mapping.

These modes are only enabled for the first greeble texture at the moment.

Moved bits around so that we can have more blending modes and a third texture.