Fix for disappearing warhead reticle in VR.
When VR is enabled, the reticle warhead may disappear if the current craft is
using custom reticles. The problem was that Reticle_7 was not being parsed from
the ini file, so the custom reticle was not identified as a CustomReticleCenter.

The fix is to add Reticle_7 as a custom reticle center.