Removed mask texture, blending modes can be specified with a string now.
There's really no need to have a mask for the greebles. The alpha channel in
each individual greeble texture is enough. That leaves open one slot for normal
mapping.

Blending modes in the MAT files can now be specified with either a digit or with
one of the following strings:

	MULTIPLY
	OVERLAY
	SCREEN
	REPLACE
	NORMAL_MAP
	UV_DISP
	UV_DISP_AND_NORMAL_MAP