Explosions in the DS2 mission now default to procedural explosions.
MechDonald's new explosions are great, but they are also a bit big and long. So,
in the DS2 mission, these explosions may obscure the tunnel. In this commit, we
check if the mission is the DS2 tunnel, and if so, we remove regular explosions
and turn them all into procedural explosions.

To turn off this feature, add the following setting to SSAO.cfg:

DS2_force_procedural_explosions = 0

Of course, setting it to 1 (or removing it), will enable it again.