Adding experimental support for 3D vision.
To enable this feature, add a file called 3DVision.cfg on your XWA root directory
with the following contents:

enable_3D_vision = 1

You'll also need to set

VR_Mode = DirectSBS
apply_lens_correction = 1

in VRParams.cfg. This is because 3D vision is currently "piggy-backing" on top
of the DirectSBS mode. So you need to activate that and disable the lens
correction normally needed for VR.