Adding InitialCockpitInstruments to prevent the Beam Weapon from exploding.
The Broken Beam Weapon Event was erroneously being triggered when a craft was
not equiped with a Beam Weapon. To prevent this, we need to get a copy of the
initial cockpit instruments and then only trigger this event if there's a beam
weapon equiped in the current craft *and* it has been damaged.