Alternate Explosions can now be loaded from ZIP files. ZIPReader is now Release version.
The alternate explosions and Dashboard lights can now be loaded from ZIP files.
Modified the code to try and load ZIP files every time a ZIP image is requested,
since only the first call will actually unzip the file (the rest of the calls
will just check for the presence of the temporary unzip directory and just
return).

Changed the ZIPReader to Release and enabled regular optimizations. This is
probably ready for testing/release now.