Force full-screen mode when 3DVision is enabled.
This commit forces full-screen mode when 3D vision is enabled.

The swap chain is released every time we switch resolutions in OnSizeChanged(),
but if we're in full-screen mode, we need to exit this mode before releasing the
swap chain.

Also, the swap chain must be initialized to nullptr before checking if we're in
full-screen mode or we'll crash when querying GetFullScreenState.