Fix the reticle's position when pose_corrected_headtracking is set.
When pose_corrected_headtracking is set, the roll is now applied in the
CockpitLook hook which is great. Unfortunately, it also messes up the position
of the reticle because XWA is not expecting roll in the camera matrix. This
commit adds a new matrix to ExternalHUDShader to fix the position of the
reticle.

We're also using a new variable that m0rgg found that gives us the vertical
offset of the reticle without capturing it. This simplifies a number of things
and makes this fix possible.