Fixed crash when DirectSBS mode was active. Added force_fullscreen.
We can't force fullscreen when 3D vision isn't active. Instead, I added a new
setting to 3Dvision called "force_fullscreen". If set, we'll force fullscreen
when in this mode.

Also: If 3Dvision is enabled, we'll automatically set VR_Mode = DirectSBS
internally, so we don't need to set this explicitly in VRConfig anymore.