New method to add 3DVision signature. Make fullscreen default.
In this commit, the 3DVision signature is pre-computed to a texture. Then,
when the texture is resized to its final dimensions, the signature is added by
copying it from this texture. This makes the process much faster. However, this
method may not work if MSAA is on.

Renamed the force_fullscreen internal parameter to make it 3DVision-specific and
made it true by default.