Fix tracking while in the hangar when pose corrected headtracking is on.
This is more of a patch rather than a real fix. If we're in the hangar, the
cockpit look hook is not active, but the new hooks are. This causes a weird
situation because the new hooks apply the tracking but SteamVR's pose is never
called. That's why tracking is broken. Here we're taking the legacy path if
pose_corrected_headtracking is ON and we're in the hangar.

This is a partial fix, the rest of the fix must be applied in the cockpit look
hook as well.

Side note: the new hooks are called even when in the hangar! That means, even
when the original cockpitlook hook is not active.