Add EffectsRenderer, re-enable Bloom from lightmaps and several other changes.
Added a new class called EffectsRenderer to implement all the changes needed for
the effects.

Added state management flags to EffectsRenderer. Most of them will have to be
checked and they will likely end up being member variables.

Applied a few changes to XwaD3dVertexShader and XwaD3dPixelShader to enable
forward-rendering stuff needed for the effects.

Moved PixelShaderCBuffer from register(b0) to register(b9) because the
D3DRendererHook is now using that slot and I didn't want to modify it.

Lightmaps are now rendered on the same draw call as the regular texture, this
means that the pixel shader receives two textures and must blend them on the
same pass (it used to be two passes before).

Several things are still broken, but we've got to start somewhere.