Fix the Hyperspace Effect.
For some reason, the hyperspace effect is super slow in the D3DRendererHook.
I've disabled rendering the original effect as that seems like a good workaround
for the slowness.

The Hyperspace effect's main problem was that the VS constant buffers were being
restored in MainSceneHook(). The fix was to avoid restoring those constants
since we'll need them later for the effect.

Another thing worth noting was that Holographic DC elements showed the
hyperspace effect through it. This was fixed by setting the right alpha blending
state in the rasterizer (keep the max alpha).