Fix the Sun Flare.
The sun flare looked weird because the color of the flare was being overwritten.
Before the D3DRendererHook, the color was set in g_ShadertoyBuffer and used
later in PrimarySurface::Flip, to render the flare. Now the order of the draw
calls is different and the color in g_ShadertoyBuffer gets overwritten with
other data. To solve this, we store the color in a new global variable and then
copy it back to g_ShadertoyBuffer, when needed.