Render lasers and projectiles at the end of the frame. Minor code refactor.
Lasers and projectiles are now rendered at the end of the frame. To do this, we
store the lasers in a list, along with associated state data. We intentionally
skip the lasers and projectiles in the CMD because they lack proper viewport
clipping.

Also, the OBJ init/close code is now in the SceneBegin and SceneEnd virtuals.