Second fix for the lasers.
When the miniature (targeted object) is about to be rendered, a hole is punched
in the depth stencil first. If we render the lasers after that hole, they will
show through. So we need to render the lasers in Execute() before we render the
HUD, and then we also need to attempt laser rendering in PrimarySurface::Flip(),
just in case nothing is targeted or there's no GUI.