Fix Shadow Mapping for the D3DRendererHook
Shadow Mapping is now working with the D3DRendererHook. The hook's WorldView
matrix is used to transform from OPT coordinates to viewspace coordinates. This
means that the ShadowMap OBJ has to be converted from OBJ coordinates back into
OPT coordinates (the Y and Z axes are swapped and the scale must be multiplied
by 40.96). The good news is that this fixed the "dancing shadows" problem at the
same time.

A small code refactor was applied too, since a lot of the back-projection code
is not used anymore. But more work is needed to remove unused variables. The
AddGeometryVertexShader is still using some of the metric reconstruction
variables that should no longer be needed with the D3DRendererHook.