Fix various artifacts.
The alpha channel for light textures was not updated properly: it was still
using the alpha from the regular texture, thus causing bright spots on areas
that shouldn't bloom.

The edge detector was executed several times per frame, depending on what was
being displayed.

The edge detector now runs when the external camera is activated.