Enable Animated Textures and a few other fixes.
The CMD in the Pilot Proving grounds was not being captured. Fixed by using the
g_isInRenderMiniature in addition to the other flags.

Lasers were still being rendered in the miniature and captured in the DC FG
buffer. Fixed by checking the type of the texture when capturing the DC buffer.
If the texture is either laser or projectile, we just skip it.

Animated Textures! The _noGlassPS and PixelShaderAnimLightMap were merged
together into a PixelShaderAnim. Seems to work, but I'm sure details are missing
and bugs will be reported.