Move procedural explosions to slot 9
Procedural explosions use slots 0 and 1. Slot 0 is for the regular texture and
slot 1 is the noise texture, but slot 1 is now used for the lightmap as well
in the D3DRendererHook. To avoid possible conflicts, let's move the noise
texture to slot 9. Procedural explosions work during regular gameplay, but the
DS2 core explosion doesn't work yet.