Fix Deferred Lasers, Laser Lights, Code cleanup.
Code cleanup. Removed RestoreSamplerState(), ComputeNormalsPass(),
SmoothNormalsPass(). Removed dead code in Direct3DDraw::Execute()

Moved grayNoiseSRV to slot 9 for procedural explosions. Sampler 9 is not
restored, but it doesn't matter because Jeremy's version doesn't use it and in
this version we just change it when needed.

Fixed deferred laser rendering. We need to save and restore the current context
when calling deferred draw calls because we might mess up the state. Also,
deferred calls must be executed right before the GUI starts, not at the
beginning of the Execute() method.

Reduced the intensity of laser lights in the cockpit. We're now using a factor
of 0.15 (it was 1 before), surely there's a better way to do this, but this
should help for now.