Fix artifacts in Hyperspace when the external camera is active.
If we jump into hyperspace while the external camera is active, the cockpit will
be rendered, but XWA will still report that the external camera is on. This
causes floating text to appear inside the cockpit. I'm adding an exception for
this edge case.

Also when jumping into hyperspace: if the cover texture is an animation with
transparency, then the background will show through. Let's mix it with a black
texture to avoid this.