Spread Vertex Buffer constants to fix the DS2 mission.
With the D3DRendererHook, register 0 is now used to hold other Vertex Buffer
constants. This seems to conflict with some VS constants we already had on the
same register. This commit spreads the usage so that we don't clobber these
values. This fixes all the glitches in the DS2 mission.

Also updated AlphaToBloomPS so that it works with the D3DRendererHook.