New Feature: The Diegetic Joystick.
The new D3DRendererHook by Jeremy gives us access to OPT coordinates and the
WorldView transform. A number of interesting applications are now possible with
this information, like being able to move meshes in response to external events.

To enable this feature, the .mat files need a new section like this:

[TEX000XX]
JoystickRoot = 0, -4.35034, -0.274126
JoystickMaxYaw = 10
JoystickMaxPitch = -8

The JoystickRoot is in meters, in the same coordinate system as the OPT, so the
Y axis is forward and Z is up. This is the location of the joint. The
JoystickMaxYaw and JoystickMaxPitch are optional and define the range of motion,
in degrees. Notice how JoystickMaxPitch is a negative number, that inverts the
pitch axis as in a regular airplane yoke.

To fully enable this feature, other changes are also needed in the joystick hook
by Jeremy: the normalized pitch and yaw must be written to a new section in the
SharedMem object. However, this can be tested by enabling joystick emulation.