Fix multiple crashes reported in TFTC hangars.
The crash was happening when the _noGlassPS shader was set, which happens when
AlphaIsntGlass = 1 is set in the material file. This shader is meant to be used
with transparent textures, but now these textures and their draw calls are
stored in a list and rendered in a deferred fashion. When storing the pixel
shader in the command list, we were using PSGetShader() to get the current
shader. Here's where this becomes tricky: PSGetShader() was probably messing up
the refcount or maybe it wasn't storing the proper pixel shader. Who knows? But
if we keep the current pixel shader, as seen by InitPixelShader(), and we keep
it as a regular ID3D11PixelShader pointer, instead of a ComPtr, then things work
just fine.