Fix the XwaD3dVRVertexShader and aspect ratio in SteamVR-Beta/
The reason we were not seeing disparity or the proper aspect ratio in the
SteamVR path was that the MainSceneHook() was calling Render() inside it and
setting the old (pancake) vertex shader + viewport in the process. We need to
override both in SteamVRRenderer::RenderScene().

Also fixed the depth buffer value in XwaD3dVRVertexShader. Otherwise we get lots
of inconsistent depth.

This commit hasn't been tested on an actual HMD yet.