Fix hangar shadows in SteamVR. Also code refactor and variable rename.
The code for hangar shadows has been commented and the relevant variables have
been renamed to make it easier to understand. A new hangar shadow vertex shader
was added for VR. Some code was refactored.

We can totally improve how shadows are rendered in this game, but that'll have
to wait for now.