Fix the Engine glow and Explosions in SteamVR mode. DATs are now fully metric.
Here we're introducing new formulas to do a metric 3D reconstruction for
DAT-based geometry, like the engine glow and the explosions. I introduced a new
VR vertex shader just for these cases since the formulas are very different from
the ones we use in SBSVertexShader.

There's also some refactoring in this commit and removal of dead code.

The key contribution is in the back-projection code for DAT geometry. It has
been verified against the regular metric data captured for regular OPTs and it
matches. The Z from this metric reconstruction is then projected using the same
code we use for OPTs and the depth is preserved by setting w = 1, to avoid
precision problems.