Add Diegetic Throttle material settings.
We can now control meshes with the throttle. Throttles can either rotate around
a cardinal axis, or slide between two points. Here's an example:

[TEX00056]
Metallic   = 0.4
Intensity  = 0.3
Glossiness = 0.08
NMIntensity = 0.4
; Rotating Throttle
ThrottleRotXAxis = -0.245698, -3.95292, 0.28396
ThrottleMaxAngle = 95

; Sliding throttle sample:
;ThrottleStart = -0.245698, -3.92522, 0.28396
;ThrottleEnd = -0.245698, -3.98063, 0.28396

We can use ThrottleRotXAxis, ThrottleRotYAxis or ThrottleRotZAxis. Angles are in
degrees and the throttle is expected to be at the "home" or "zero position" in
the OPT.