Add normals to the OBJ dumps. Vertex hash map for normals smoothing.
We really can't smooth the normals in this path. We only see a subset of all the
vertices so there are many normals we can't smooth. We either need full
visibility of the OPT, or we need to do it outside of ddraw.

Still, it's good to have some reference code to do a vertex hash map, in case we
need it at some point.