Fade lights that don't come from Suns. Code cleanup.
XWA puts lights on planets as well as suns. However, some people have pointed
out that this looks like a bug. So let's fade the lights that come from planet
objects and see how that goes. This behavior can be reverted by adding the
following line in SSAO.cfg:

	min_light_intensity_after_fade = 1

We don't need the VRRenderer after all, so let's remove it.

Removed the code that smoothed the normals. It wasn't complete and XwaSmoother
now does the job right.