Use shadow mapping in the hangar. Add support for per-mesh AABBs.
Hangar shadows needs to be made more efficient in VR and a number of glitches
need to be fixed. Both goals can be achieved if we switch to using shadow
mapping for the hangar.

There are still some artifacts in this code, so that's why it's disabled by
default for now.

Also, added per-mesh AABBs. This will be useful later.