Fix artifacts in the hangar shadows. Fix diegetic throttle.
If there's no shadowmap OBJ for the cockpit when in the hangar, some artifacts
will appear because we're forcing 2 lights in this situation. We need to disable
the first light in this case (by setting the black level to 1).

If there's a throttle and a hyper-throttle, the cached transform for one will
override the other. Let's disable the cache for now.