Fix shadow artifacts in the hangar, add ranges, and apply new shadows in VR path
Each shadow map now has a min/max range which also helps compute the proper
bias for it.

We shouldn't fade lights while in the hangar. The illumination doesn't match the
suns anyway and the hangar just looks dark for no good reason.

The VR path is now using the soft hangar shadows, so the performance should be
much better.

Added a toggle in DDraw.cfg for the soft hangar shadows:

	EnableSoftHangarShadows

It's set to 1 by default.

These changes could be a prelude to enabling Cascaded Shadow Maps for global
shadows while flying.