Restore DAT animations in Direct3DDevice::Execute()
I removed the animations path from Execute() a little too quickly. This restores
DAT animations, but it may need a bit more work.

Also, adding some code to increase the refcount of cached SRVs, just to be on
the safe side (that way they won't be released). SRVs should be released in
DeviceResources::ResetExtraTextures() with a loop that checks until the refcount
reaches 0.