Fix Sun Flare by disabling backface culling. Restore _extraTextures Release()
The sun flare was not showing because backface culling was enabled in a previous
commit. To solve this, we need to set the original rasterizer state from
DeviceResources (that rasterizer doesn't cull backfaces).

DeviceResources::ResetExtraTextures() crashed again. So, reverting to the
previous code, even if it might leak some textures. I really need to fix this,
but not sure how.