Set backface culling to off by default. Restore transparency during hyperspace
Turns out we don't need to blend the background of transparent textures with
black anymore. In fact, doing so, will obscure the cockpit glass if it's
damaged.

Enabling backface culling will show some gaps in SteamVR for some cockpits. So,
for now, we need to disabling culling, but this time I've added a DDraw setting
to enable it if necessary: CullBackFaces.