Fix crash in the Tech Room when SteamVR is used.
When the Tech Room is displayed, we intentionally reset the meshes each time a
new craft is loaded. However, we should not do that at the end of the frame,
we should to it at the beginning, right after the cached _lastVertexMesh,
_lastVertexMeshView, etc pointers are set to nullptr. This was causing SteamVR
to crash (but not pancake mode, for some reason).

Also changed the meaning of the g_bResetCachedMeshes flag. It's now a request to
reset the meshes on the next frame.